Complete VulkanManager implementation with full production features
- Add complete Vulkan function pointer loading for all core functions
- Implement device enumeration and selection
- Add pipeline creation with SPIR-V shader loading
- Implement buffer copy and pipeline barrier commands
- Add memory size queries and usage tracking
- Complete descriptor set management
- Add proper error handling and logging throughout
- Implement waitForCompletion with timeout parameter

Production-ready Vulkan compute backend for RawrXD Agentic framework.

+++ src/security/SecurityHardening.cpp
@@ -68,7 +68,7 @@ std::string AuditEvent::ToJson() const {
 
 std::string AuditEvent::ComputeHash() const {
     std::stringstream ss;
-    ss << eventId << type << timestamp << agentId << intentId
+    ss << eventId << static_cast<int>(type) << timestamp << agentId << intentId
        << action << target << details << previousEventId;
     return ComputeSHA256(ss.str());
 }
